% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_intersect.R
\name{date_intersect}
\alias{date_intersect}
\alias{int_intersect}
\alias{dtime_intersect}
\title{date_intersect}
\usage{
date_intersect(date_x, date_y)

int_intersect(int_x, int_y)

dtime_intersect(x, y)
}
\arguments{
\item{date_x}{date vector, [date_begin, date_end]}

\item{date_y}{same as date_x}

\item{int_x}{interval x}

\item{int_y}{interval y}

\item{x}{dtime object}

\item{y}{same as x}
}
\description{
Intersect of two date interval
}
