% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st-get_moveInfo.R, R/st-revise_locaiton_multi.R
\name{distToCentralPeriod}
\alias{distToCentralPeriod}
\title{The distance of locations from the center}
\usage{
distToCentralPeriod(lon, lat, n_period)

distToCentralPeriod(lon, lat, n_period)
}
\description{
The distance of locations from the center

The distance of locations from the center
}
