% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtime-S3Class.R
\name{dtime}
\alias{dtime}
\title{dtime}
\usage{
dtime(
  data = NULL,
  station = NULL,
  begin = Sys.Date(),
  end = Sys.Date(),
  by = "day"
)
}
\arguments{
\item{data}{numeric vector or matrix}

\item{station}{site id or name}

\item{begin}{begin date}

\item{end}{end date}

\item{by}{increment of the sequence}
}
\description{
dtime
}
