% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_meteV3.R
\name{flag_realvalue}
\alias{flag_realvalue}
\title{flag_realvalue}
\usage{
flag_realvalue(x, values)
}
\arguments{
\item{x}{numeric or matrix}

\item{values}{Should be in descend order}
}
\description{
flag_realvalue
}
