% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtime-S3Class.R
\name{Interp}
\alias{Interp}
\title{Interp}
\usage{
Interp(data_in, dist, smax = 100, nmax = 10, maxgap = 5, time_day,
  stationId, sinkfile1 = NULL, sinkfile2 = NULL, ...)
}
\arguments{
\item{data_in}{matrix data with dim(date*station)}

\item{smax}{maximum searching radius of linear interpolation (unit:km), default is 100km.}

\item{nmax}{maximum number stations used in linear interpolation, default is 10.}

\item{maxgap}{maximum gap used in naApprox interpolation}

\item{sinkfile}{file that used to save linear interpolation information}
}
\description{
main funtion of interpolation, combined with na.approx, linear, hisavg interpolations
}

