% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interp_approx.R
\name{interp_approx}
\alias{interp_approx}
\title{interp_approx}
\usage{
interp_approx(xx, stationInfo, maxgap = 5, verbose = TRUE,
  .parallel = FALSE, ...)
}
\arguments{
\item{xx}{list of dtime objects which have missing values}

\item{stationInfo}{stationInfo object returned by Missinfo function}

\item{maxgap}{maximum number of consecutive \code{NA}s to
    fill. Any longer gaps will be left unchanged. Note that all methods listed
    above can accept \code{maxgap} as it is ultimately passed to the 
    \code{default} method.}

\item{verbose}{boolean whether print progrss message?}

\item{.parallel}{boolean}

\item{...}{ignored}
}
\description{
Linear interpolation with nearest values
}
