% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interp_hisavg.R
\name{interp_hisavg}
\alias{interp_hisavg}
\title{interp_hisavg}
\usage{
interp_hisavg(xx, stationInfo, ...)
}
\arguments{
\item{xx}{list of dtime objects which have missing values}

\item{stationInfo}{stationInfo object returned by Missinfo function}

\item{...}{ignored}
}
\description{
interp_hisavg
}
