% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interp_main.R
\name{interp_main}
\alias{interp_main}
\title{interp_main}
\usage{
interp_main(
  df,
  st,
  smax = 100,
  nmax = 10,
  maxgap = 5,
  verbose = TRUE,
  sinkfile1 = "interp_lm1.log",
  sinkfile2 = "interp_lm2.log",
  ...
)
}
\arguments{
\item{df}{data.frame, with dimension of [n_date, 1+n_station], and the first
column is date.}

\item{st}{data.frame, with column at least of 'site', 'lat', 'lon' (degree).}

\item{smax}{maximum searching radius (km) of linear interpolation, default is 100km.}

\item{nmax}{maximum number stations used in linear interpolation, default is 10.}

\item{maxgap}{maximum gap used in interp_approx interpolation}

\item{...}{ignored}
}
\description{
Main funtion of interpolation, combined with na.approx, linear
and hisavg interpolation.
}
\examples{
\dontrun{
data("prcp")
data("st840")
r <- interp_main(prcp, st840, smax=200)
}
}
