% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearInterp.R
\name{linearInterp}
\alias{linearInterp}
\title{linearInterp}
\usage{
linearInterp(xtrim, stationInfo, dist, nmax = 10, sinkfile = NULL)
}
\arguments{
\item{x:}{dtime class whose data have missing values}

\item{StationInfo:}{single one stationInfo returned by Missinfo function}

\item{dist:}{Station distance from each other dim = (n * n)}
}
\description{
linearInterp
}
