% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st-revise_locaiton_multi.R
\name{locMatchedScore}
\alias{locMatchedScore}
\title{location matched score}
\usage{
locMatchedScore(x, y)
}
\value{
- `3`: all of `lon`, `lat` and `alt` matched
- `2`: two of `lon`, `lat` and `alt` matched
- `1`: one of `lon`, `lat` and `alt` matched
}
\description{
location matched score
}
