% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtime-S3Class.R
\name{MissInfo.matrix}
\alias{MissInfo.matrix}
\title{MissInfo}
\usage{
\method{MissInfo}{matrix}(data, time_day = as.Date(rownames(data)),
  stationId = colnames(data), clipdata = FALSE, progress = TRUE, ...)
}
\arguments{
\item{data}{matrix data with row dimemsion is date and col is stations}

\item{time_day}{data's responding daily date}

\item{stationId}{stations number id or names}

\item{clipdata}{if clipdata is true, NAs at head and tail will be trimed. And trimed data will be return}
}
\description{
detect all station missing infomation for single meteorological element
}

