% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missInfo-package.R, R/missInfo.R
\docType{package}
\name{MissInfo}
\alias{MissInfo}
\alias{_PACKAGE}
\alias{MissInfo-package}
\alias{MissInfo.default}
\alias{MissInfo.dtime}
\alias{MissInfo.matrix}
\alias{MissInfo.list}
\title{MissInfo}
\usage{
MissInfo(data, clipdata = FALSE, ...)

\method{MissInfo}{default}(x, station, time_day, clipdata = FALSE,
  detailInfo = TRUE, daysInfo = TRUE, collapse = ", ")

\method{MissInfo}{dtime}(x, ...)

\method{MissInfo}{matrix}(data, time_day = as.Date(rownames(data)),
  stationId = colnames(data), clipdata = FALSE, progress = TRUE, ...)

\method{MissInfo}{list}(xlist, clipdata = FALSE, .progress = "text",
  .parallel = FALSE, ...)
}
\arguments{
\item{data}{matrix data with row dimemsion is date and col is stations}

\item{clipdata}{If clipdata is true, NAs at head and tail will be trimed. 
And trimed data will be return.}

\item{x}{one station meteorological one column matrix data. Rownames is daily time string}

\item{time_day}{data's responding daily date}

\item{detailInfo}{If true detail missing information will be return, else missing information is na.}

\item{daysInfo}{If true missing seg days will be return}

\item{collapse}{missinfo info collapse}

\item{stationId}{stations number id or names}
}
\description{
detect missing information of meteorological data and interpolate missing values using common interpolation methods

detect one station missing infomation for single meteorological element
}
\details{
If all(is.na(x)) is true, then return a dtime begin and end date is 1951-01-01
}
\author{
\strong{Maintainer}: Dongdong Kong \email{kongdd.sysu@gmail.com}

}
\keyword{information,}
\keyword{internal}
\keyword{interpolate}
\keyword{missing}
