% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtime-S3Class.R
\name{MissInfoSingle}
\alias{MissInfoSingle}
\title{MissInfo.Single}
\usage{
MissInfoSingle(x, station, time_day, clipdata = FALSE, detailInfo = TRUE,
  daysInfo = TRUE, collapse = ", ")
}
\arguments{
\item{detailInfo}{If true detail missing information will be return, else missing information is na.}

\item{daysInfo}{If true missing seg days will be return}

\item{collapse}{missinfo info collapse}

\item{x:}{one station meteorological one column matrix data. Rownames is daily time string}

\item{clipdata:}{return trim head and tail blank data}
}
\description{
detect one station missing infomation for single meteorological element
}
\details{
if all(is.na(x)) is true, then return a dtime begin and end date is 1951-01-01
}

