% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interp_approx.R
\name{na_approx.dtime}
\alias{na_approx.dtime}
\title{na_approx.dtime}
\usage{
na_approx.dtime(x, maxgap = 5, ...)
}
\arguments{
\item{x}{dtime object}

\item{maxgap}{maximum number of consecutive \code{NA}s to
    fill. Any longer gaps will be left unchanged. Note that all methods listed
    above can accept \code{maxgap} as it is ultimately passed to the 
    \code{default} method.}

\item{...}{ignored}
}
\description{
na_approx.dtime
}
