% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interp_hisavg.R
\name{na_hisavg.dtime}
\alias{na_hisavg.dtime}
\title{na_hisavg.dtime}
\usage{
na_hisavg.dtime(x, ...)
}
\arguments{
\item{x}{dtime class}
}
\description{
using history average y_his replace missing values
}
