% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_var.R
\name{read_var}
\alias{read_var}
\title{read files of China Meteorological data V3}
\usage{
read_var(indir, varname = NULL, drop = NULL, limits = NULL)
}
\arguments{
\item{indir}{Directory of txt files}

\item{varname}{meteorological variable names}

\item{drop}{Numeric, which variable to be droped}

\item{limits}{Numeric, If not null, only read the first \code{limits} files.}
}
\description{
read files of China Meteorological data V3
}
