% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_data.R
\docType{data}
\name{st_met2176_hourly}
\alias{st_met2176_hourly}
\title{station information of 2176 hourly meteorological data}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 2176 rows and 9 columns.
}
\usage{
data(st_met2176_hourly)
}
\description{
station information of 2176 hourly meteorological data
}
\references{
1. https://data.cma.cn/data/detail/dataCode/A.0012.0001.html
}
\keyword{datasets}
