% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_data.R
\docType{data}
\name{st_met2370}
\alias{st_met2370}
\title{station information of 2370 daily meteorological data from Ren GuoYu}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 2370 rows and 6 columns.
}
\usage{
data(st_met2176_hourly)
}
\description{
station information of 2370 daily meteorological data from Ren GuoYu
}
\keyword{datasets}
