% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_data.R
\docType{data}
\name{st_met2481}
\alias{st_met2481}
\title{station information of 2481 daily meteorological data from Gu Xihui}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 2481 rows and 13 columns.
}
\usage{
data(st_met2481)
}
\description{
station information of 2481 daily meteorological data from Gu Xihui
}
\keyword{datasets}
