% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_meteV3.R
\name{tidy_meteV3}
\alias{tidy_meteV3}
\title{handle of meteorological flags}
\usage{
tidy_meteV3(df)
}
\arguments{
\item{df}{data.frame}
}
\description{
handle of meteorological flags
}
\examples{
data("mete840")
df1 <- tidy_meteV3(mete840)
df2 <- tidy_meteV3(mete840[, 1:10])
}
