% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zip_dates.R
\name{zip_dates}
\alias{zip_dates}
\title{zip_dates}
\usage{
zip_dates(dates, by = NULL, format = NULL, check_duplicate = FALSE)
}
\arguments{
\item{dates}{Date vector}

\item{by}{increment of the sequence, could be one of 'day', 'month' and 'year'.}

\item{format}{character string.}

\item{check_duplicate}{If true, duplicated formated dates by `format` will be
removed. If by = 'month' or 'year', suggest setting \code{check_duplicate} as 
TRUE.}
}
\description{
zip_dates
}
\examples{
library(lubridate)
data("station_HisPos")
d <- station_HisPos[site == 50136, ]
dates <- d[, make_date(year, month, day)]

s1 <- zip_dates(dates)
s2 <- zip_dates(dates, by = "year", check_duplicate = TRUE)
}
