% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_meteV3.R
\name{cal_exceed}
\alias{cal_exceed}
\alias{cal_below}
\title{cal_exceed}
\usage{
cal_exceed(x, values)

cal_below(x, values)
}
\arguments{
\item{x}{numeric or matrix}

\item{values}{Should be in descend order}
}
\description{
cal_exceed
}
