% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interp.R
\name{Interp}
\alias{Interp}
\title{Interp}
\usage{
Interp(xlist, stationInfo, dist, nmax = 10, maxgap = 5,
  sinkfile = NULL, .parallel = T, ...)
}
\arguments{
\item{nmax}{maximum number stations used in linear interpolation, default is 10.}

\item{maxgap}{maximum gap used in naApprox interpolation}

\item{sinkfile}{file that used to save linear interpolation information}

\item{data_in}{matrix data with dim(date*station)}

\item{smax}{maximum searching radius of linear interpolation (unit:km), default is 100km.}
}
\description{
Main funtion of interpolation, combined with na.approx, linear 
and hisavg interpolations
}
