% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missInfo-package.R
\docType{package}
\name{MissInfo-package}
\alias{missInfo}
\alias{missInfo-package}
\title{MissInfo}
\description{
detect missing information of meteorological data and interpolate missing values using common interpolation methods
}
\author{
\strong{Maintainer}: Dongdong Kong \email{kongdd.sysu@gmail.com}

}
\keyword{information,}
\keyword{internal}
\keyword{interpolate}
\keyword{missing}
