% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missInfo.R, R/missInfo.matrix.R
\name{missInfo}
\alias{missInfo}
\alias{missInfo.matrix}
\alias{missInfo.data.frame}
\alias{missInfo.list}
\title{missInfo}
\usage{
missInfo(data, date, station, clipdata = FALSE, ...)

\method{missInfo}{matrix}(
  mat,
  date = as.Date(rownames(mat)),
  site = colnames(mat),
  clipdata = FALSE,
  verbose = TRUE,
  .parallel = FALSE,
  ...
)

\method{missInfo}{data.frame}(df, clipdata = FALSE, verbose = TRUE, .parallel = FALSE, ...)

\method{missInfo}{list}(lst, clipdata = FALSE, verbose = TRUE, .parallel = FALSE, ...)
}
\arguments{
\item{...}{ignored}

\item{mat}{matrix data with row dimemsion is date and col is stations}

\item{df}{data.frame, with dimension of [n_date, 1+n_station], and the first
column is date.}

\item{lst}{list of dtime object}
}
\description{
missInfo
}
\examples{
data('prcp')
time_day <- prcp$date
site <- colnames(prcp)[-1]
mat  <- as.matrix(prcp[, -1])
Info1 <- missInfo:::MissInfo.matrix(mat, clipdata = TRUE, time_day, site, verbose = TRUE)
Info2 <- MissInfo(prcp, clipdata = TRUE, time_day, site, verbose = TRUE)
all.equal(Info2, Info1)
}
