% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MissInfo.R, R/MissInfo.matrix.R
\name{missInfo}
\alias{missInfo}
\alias{missInfo.default}
\alias{missInfo.dtime}
\alias{missInfo.matrix}
\alias{missInfo.data.frame}
\alias{missInfo.list}
\title{missInfo}
\usage{
missInfo(data, date, station, clipdata = FALSE, ...)

\method{missInfo}{default}(
  x,
  date,
  station = "S1",
  clipdata = FALSE,
  Info_detail = TRUE,
  Info_days = TRUE,
  collapse = ", "
)

\method{missInfo}{dtime}(x, ...)

\method{missInfo}{matrix}(
  mat,
  date = as.Date(rownames(mat)),
  site = colnames(mat),
  clipdata = FALSE,
  verbose = TRUE,
  .parallel = FALSE,
  ...
)

\method{missInfo}{data.frame}(df, clipdata = FALSE, verbose = TRUE, .parallel = FALSE, ...)

\method{missInfo}{list}(lst, clipdata = FALSE, verbose = TRUE, .parallel = FALSE, ...)
}
\arguments{
\item{date}{corresponding date of x (support monthly or daily)}

\item{station}{character}

\item{clipdata}{If clipdata is true, NAs at head and tail will be trimed.
And trimed data will be return.}

\item{...}{ignored}

\item{x}{one station meteorological one column matrix data. Rownames is daily time string}

\item{Info_detail}{If true detail n_miss information will be return, else
n_miss information is na.}

\item{Info_days}{If true n_miss seg days will be return}

\item{collapse}{missinfo info collapse}

\item{mat}{matrix data with row dimemsion is date and col is stations}

\item{df}{data.frame, with dimension of [n_date, 1+n_station], and the first
column is date.}

\item{lst}{list of dtime object}
}
\description{
detect one station n_miss infomation for single meteorological element
}
\details{
If all(is.na(x)) is true, then return a dtime begin and end date is 1951-01-01
}
\examples{
data('prcp')
time_day <- prcp$date
site <- colnames(prcp)[-1]
mat  <- as.matrix(prcp[, -1])
Info1 <- missInfo:::MissInfo.matrix(mat, clipdata = TRUE, time_day, site, verbose = TRUE)
Info2 <- MissInfo(prcp, clipdata = TRUE, time_day, site, verbose = TRUE)
all.equal(Info2, Info1)
}
