% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missInfo-package.R, R/missInfo.R
\docType{package}
\name{MissInfo}
\alias{MissInfo}
\alias{_PACKAGE}
\alias{MissInfo-package}
\alias{missInfo.default}
\alias{missInfo.dtime}
\title{MissInfo}
\usage{
\method{missInfo}{default}(
  x,
  date,
  station = "S1",
  clipdata = FALSE,
  Info_detail = TRUE,
  Info_days = TRUE,
  collapse = ", "
)

\method{missInfo}{dtime}(x, ...)
}
\arguments{
\item{x}{one station meteorological one column matrix data. Rownames is daily time string}

\item{date}{corresponding date of x (support monthly or daily)}

\item{station}{character}

\item{clipdata}{If clipdata is true, NAs at head and tail will be trimed.
And trimed data will be return.}

\item{Info_detail}{If true detail n_miss information will be return, else
n_miss information is na.}

\item{Info_days}{If true n_miss seg days will be return}

\item{collapse}{missinfo info collapse}
}
\description{
detect missing information of meteorological data and interpolate missing values using common interpolation methods

detect one station n_miss infomation for single meteorological element
}
\details{
If all(is.na(x)) is true, then return a dtime begin and end date is 1951-01-01
}
\author{
\strong{Maintainer}: Dongdong Kong \email{kongdd.sysu@gmail.com}

}
\keyword{information,}
\keyword{internal}
\keyword{interpolate}
\keyword{missing}
